Add initial conductor API and service
Add an initial conductor API and service.  This service allows
operation on the four object types -> bays, pods, services, and
containers and allows the ability to list, read, write, or delete
those objects in the database.  The implementation of the
list/read/write/delete is incomplete and should come in a follow-up
patch.

Change-Id: I60e9070e4b5aeaeddba67233b99dd0e3a3cffe22