Containers Service Spec
Containers share many features in common with Nova instances. For the common
features, virt drivers for Nova can be used to surface basic instance
functionality. For features that go beyond what can be naturally fit within
a virt driver, we propose a new API service that allows for advanced features
to be added without conflating the worlds of instances and containers.

Re-post of the spec originally posted at:
https://review.openstack.org/114044

Change-Id: Ic5d95f5d6ba0d6845023bf59f9bad4377d50578f