Add image_id and node_count to bay
The idea of bay-create is that you create a grouping of machines that run
container software.  In order to do this, an image_id stored in glance is
necessary to pull the base image to modify for launching by either heat or
Ironic.  A node_count is needed as well, to identify how many ironic nodes
you want to launch.

Change-Id: I5689d024f4ece93369498bd7f64cde0b8a59e292