Add unit tests for the conductor AMQP API
This tests create and delete at the AMQP level and makes certain the AMQP
code is being called.  This does not actually test the connection to the
handlers, which would be more like a functional test.

This test case purpose is meant to introduce pain when changing the
interface.  The interface should not be changed, a new version should
be implemented.

Change-Id: Icfca04df842497515c20138ba58694805b266b55