Remove automatic deletion of failed stacks
Prior to this patch, Magnum would delete stacks that entered the
FAILED state.  This resulted in the information about why the
orchestration failed being lost to the user, and only avaialble
in the system logs.

Instead, keep the stack around so it can be inspected and delete
it when the bay is deleted.

Change-Id: Ie61b84ee00ea4f96580f5725e72ca35b855e0417
Closes-Bug: 1412213