Create heat template for docker-swarm bay
This template will spin up a single swarm-manager node and N number of
swarm-agent nodes. Every node, including the manager, is running a
Docker daemon and a swarm-agent service advertising the docker daemon.
The swarm-manager node is running an additional swarm-manager service
listening on port 2376. The swarm-manager service acts a a single
docker api for interacting with the entire cluster.

This template was adapted from the existing heat-kubernetes bay template.
As such, it uses identical neutron network topology.

Implements: blueprint docker-bay-heat-template

Change-Id: I22276022bbb4155ccdce0f75c47d46492a5ac821