Add status attribute to bay
When a Bay is created, we use Heat to bring up resources to use for the Bay.
We would like to see the related status in the Bay resource that tracks what's
happening. We use the new status attribute to show what is happening with Heat
behind the scenes.

Partially-Implements: blueprint magnum-bay-status

Change-Id: I7b841f2afbe575b039ebbde0e19d8fc464246608