Change replicas to 2 in dev quick start
As the dev quick start will only create two minions, so if when
create 4 replicas, there will be only 2 pods create succeed and
the other 2 pods will be in Pending status.

This patch is updating the replicas to 2 to make sure all pods
in the replication controller can be created successfully.

Change-Id: Ib9373487e76d4296dc412cc8a4f61af1bca94647