Allow pod delete to succeed when not found on bay
If a pod is deleted from kubernetes outside of magnum, and then
a pod-delete is issued through magnum, the delete will succeed
but will leave the record in the database. This change raises
an exception when the pod is not found, but catches it in the
conductor and proceeds to delete the record from the database.

Change-Id: I83a6b3e68eef434fb7d1b4cba92acb17abc6bc4d
Closes-bug: #1412587