Change link of quick start to git.openstack.org
The current start guide is using github.com for the reference of
the quick start guide, but this might not always the latest quick
start.

This patch is using git.openstack.org as the dev quick start guide
which can make sure we can always get the latest quick start.

Change-Id: Ie0902166a3e5280f42d5479db10026528c548a68