Add master flavor
Allow specification of master flavor. This allows choice of flavor for
the master node and the flavor size of the nodes that the containers
will run on. Often you will want a large flavor size for the nodes
running docker to load many containers, but a small flavor size for the
master just running control plane services for the docker nodes.

Change-Id: Id7691cde80ea403cfe38634958bb009a24e017dc
Partial-Bug: 1418713