Add devstack module to contrib
Add devstack module and with instructions on using it included in the
quickstart guide. The old manual process for using magnum with devstack
has been kept to retain information on how to deploy magnum manually. A
localrc for use with devstack has also been included rather than forcing
the user to copy and paste from the docs.

Closes-bug: 1430053
Change-Id: I373021378fd25d5a6cdb6213eaa21f8045a66be6