Allow bay resource management by "name".
Currently all the magnum APIs can only be called
with resource id. This patch supports API calling
with name of the resources along with the id also.

Eg.
	magnum bay-delete bay1
This will delete a bay with name bay1.

This patch supports named based resource management
for bay resource and below is the list of APIs:
	GET /v1/bays/<name/id>
	DELETE /v1/bays/<name/id>

Partially implements blueprint name-based-resource-management
Change-Id: I5e5b849aac4929237d50eac2a2e889c9f94be945