Rename bay's master_address to api_address
The idea of having a master is a Kubernetes concept and isn't
necessarily a bay concept. In Magnum, we are actually using this
value as the API endpoint. Thus, this patch renames the column
in order to be more clear about what this value is used for, and
also to be more agnostic across each Container Orchestration
Environment.

Partial-Implements: blueprint multiple-bay-templates

Change-Id: Ifaaf626377b8b302e5b390deea53be7be85e88f7