Convert to new DevStack plugin model for functional testing
Switch over to the new plugin style as documented here:
http://docs.openstack.org/developer/devstack/plugins.html

Makes it very easy to setup a CI job. Added a new
directory for functional job and the tests are run
from the post_test_hook.sh. There's even a "magnum bay-list"
in the shell script to show that we are actually making
calls to the magnum API endpoint and things are working
fine. Note that this is just the minimal plumbing needed
and there's a lot more effort needed to come up with
proper test cases for testing magnum functionality.

Closes-Bug: #1439621
Change-Id: I88d45015338ef55713b32de196f9071c8fc2031b