Destroy the related resources when delete a bay
Currently, when we delete a bay, we will delete the stack firstly,
then try to detroy the bay record in db. And at this time, a
BayNotEmpty exception may raised if some of pods, services or rc
are related to this bay. In the meanwhile, the stack should be already
deleted, actually, the left bay record is unavailable.

This patch destroy the related resources when delete a bay.

Change-Id: I56cd8813417bd4a6f6c340deb8afd88d5516f225
Closes-bug: #1440630