Add command field for container
Currently, We create a container, if the "Cmd" attribute of the
corresponding image is not null, the container can be created
successfully, but it will result in a failure if the "Cmd" attribute
is null.

This patch add a new column 'command' for container, and user can
choose to pass it when creating a container.

Closes-bug: #1445504

Change-Id: Icdd47b2343fd1ebc6ebc376764cedd0340407721