Add support for API microversions
This patch adds support in the REST API to accept a new
X-OpenStack-Magnum-API-Version header. This header is also returned
to clients as a hint to improve client response handling.

Additionally, new Minimum-Version and Maximum-Version headers
are also now returned, so as to inform clients of the supported range.

The requested version is stashed in the pecan.request.version object.

For reference as to why this approach was chosen, see the Nova spec on
microversioning:
http://specs.openstack.org/openstack/nova-specs/specs/kilo/approved/api-microversions.html
Ironic:
https://review.openstack.org/#/c/150821

Implements: blueprint api-microversions
https://blueprints.launchpad.net/magnum/+spec/api-microversions

Co-Authored-By: ShaoHe Feng <shaohe.feng@intel.com>

Change-Id: I795f83cd829098e3d1f08d89ea33a8d986797d9f