Web Interface for Magnum in Horizon
Currently there is no way for a user to interact with Magnum through a web based
user interface, as they are used to doing with other OpenStack components. This
implementation aims to introduce this interface as an extension of Horizon (the
OpenStack Dashboard) and expose all the features of Magnum in a way familiar to
users.

Implements: blueprint magnum-horizon-plugin

Change-Id: I832ee1fdf0c9b2f1ece8b01329d5c7f5bec8ae3c