Initial Heat template for Mesos
The template provision a Mesos cluster with Marathon and Docker.
It is assumed that users provide a Ubuntu base image with Mesos
pre-installed (how to build such image will be documented in another
patch). A provisioned Mesos cluster consists of two types of nodes:
master and slave.

A master node will have the following processes running:
* Zookeeper: A distributed configuration and service discovery service.
* Mesos-master: A process for orchestrating the running of tasks on slaves.
* Marathon: A cluster-wide init and control system for services.

A slave node will have the following processes running:
* Docker: The container engine.
* Mesos-slave: A process for running tasks.

The following features are NOT yet supported:
* Docker storage support.
* Docker network (Flannel or Neutron support).
* Multiple Mesos master nodes.
* Additional Mesos frameworks (i.e. Chronos, Aurora, etc.).
Some of them will be addressed in later patches. Some of them need further
discussion.

Change-Id: Ibf1be3cfb44290d64798865cd18e47c95d937e93
Partial-Implements: blueprint mesos-bay-type