Configure Fedora Atomic for Kubernetes load balancer feature
Kubernetes service provides the LoadBalancer feature

This is supported through an OpenStack plugin in Kubernetes code.
Enabling this feature requires configuring the cluster properly for
Kubernetes to interface with Neutron.  Kubernetes backend will then
create the Neutron load balancer pool, members, VIP and monitor,
and manage the pool members as pods are added/removed from the
service.

This patch updates the Fedora Atomic heat templates and scripts for
the following changes:

1. Update command line parameters for Kubernetes services.

2. Pass along OpenStack info and generate the configuration file
for Kubernetes to talk to OpenStack.

3. Adjust the name for minion and private network to match what
Kubernetes expects.

4. Let kubelet register the minion automatically instead of
registering manually.

With this patch, a Kubernetes service containing the attribute
"type: LoadBalancer" will have a load balancer created in Neutron.
The new image fedora-21-atomic-6.qcow2 needs to be used since it
contains required fixes for the kube-controller-manager.
The load balancer support will be disabled by default, the user
needs to log into the master node to enter the password to
enable the feature.  See the guide for details.

The templates for CoreOS and Ironic are in following patches.

Partially-Implements: blueprint external-lb
Change-Id: I2c4b854f94fb2dfc99fc460c897f33bf59d260e3