Configure CoreOS for Kubernetes load balancer feature
Kubernetes service provides the LoadBalancer feature

This is supported through an OpenStack plugin in Kubernetes code.
Enabling this feature requires configuring the cluster properly for
Kubernetes to interface with Neutron.  Kubernetes backend will then
create the Neutron load balancer pool, members, VIP and monitor,
and manage the pool members as pods are added/removed from the
service.

This patch updates the CoreOS heat templates for host name and
network name that match what Kubernetes code expects.
The scripts used by these templates are in the patch for Fedora Atomic
which this patch has a dependency on.

Partially-Implements: blueprint external-lb
Change-Id: I9b927624968a463625f399e0fc5deec2c6173e5f