Configure Ironic for Kubernetes load balancer feature
Kubernetes service provides the LoadBalancer feature

This is supported through an OpenStack plugin in Kubernetes code.
Enabling this features require configuring the cluster properly for
Kubernetes to interface with Neutron.  Kubernetes backend will then
create the Neutron load balancer pool, members, VIP and monitor,
and manage the pool members as pods are added/removed from the
service.

This patch updates the Ironic heat templates to adjust the minion
name and private network name to match what Kubernetes expects.

Additional scripts are from the patch for Fedora Atomic which
this patch has a dependency on.

Partially-Implements: blueprint external-lb
Change-Id: I74eaffe280b8ed20fce432d92fa6fc29048de8a5