Rename image_id to image in container object
Currently, magnum used "image_id" to reflect the image in container,
but actually, we can support use both image name and image id to create
a container. The name "image_id" may confuse user that we only support
"image_id" when create a container.

This patch rename the "image_id" field to "image" in container object.

Change-Id: Ie3639feb2a2a90300662b74f186ef998d38d6cc4
Closes-Bug: #1466278
Closes-Bug: #1465558