Make swarm work with atomic image
The docker swarm latest image need docker engine 1.6.0 above.
Our atomic image is old which integrated with 1.5.0.
Usually we'd like a new image with docker 1.6.0, but fedora 21 docker
package seems have issue with docker 1.6.0 check bug in:
https://github.com/docker/swarm/issues/781

In order to not block our magnum work, we'd like to bind to
swarm 0.2.0, as that image worked.If fedora 21 image can work
with docker 1.6.0, we would upgrade our new image. So it is
a workaround here.

Closes-Bug: #1466817
Change-Id: I048da03ca64a57bffb3b2872009855110fe45bd5