Instruction for building Fedora Atomic image
For Magnum development, we use a Fedora Atomic image prebuilt with a certain
version of Docker, Kubernetes, etcd and flannel.  The image does need to be
updated from time to time, or a user might want to test a specific version.
This document contains the instruction to build the image.
Optionally, you can also update the existing image.

Co-Authored-By: Steven Dake <stdake@cisco.com>
Change-Id: I62003abfa9077dad07988ecbce8fc5bd566f2299