Implement bay smart scale down
We currently allow Magnum to scale down bays by removing nodes from
the Bay's ResourceGroup by updating the heat stack that created the bay.
The problem with this approach is that Heat decides which node to delete,
and all containers on that node will also be deleted. This patch make
the selection process smarter: we ask Heat to delete Bay node(s)
that have NO CONTAINERS on them.

Change-Id: Ic4d754dfe358dd8aad020bc3b80a4c9f3f1525fc
Implements: blueprint magnum-smart-bay-scale-down