Docker container-create fails with Unicode is not supported
Problem description:
If a docker swarm baymodel, docker swarm bay, and docker container are
created according to instructions listed in the Magnum quick-start guide
(magnum/doc/source/dev/dev-quickstart.rst), then the container-create
operation fails with the following error:

   ERROR: Docker internal Error: decoding Unicode is not supported (HTTP 500)

The root cause is that the container image name is being passed as a
unicode value (e.g. u'cirros') to the docker conductor handler's
container_create() method, and this in turn is passed to _encode_utf8().
In _encode_utf8(), it's assumed that a string is being passed in, so a
conversion to unicode is attempted by passing the (unicode) image name
to six.u(). However, six.u() does not support unicode conversion of
unicode objects.

Fix description:
Make the _encode_utf8() method more resilient by allowing both string and
unicode arguments. Only do the intermediate conversion to unicode for
string arguments.

Change-Id: Ie60099172c4b6fa83c0d5d10aba0f9c24a349106
Closes-Bug: #1471984