Add return value to mocks to allow serialisation
In the container tests the mocks being passed in are being serialised
by pecan however they have no return value and therefore cannot be
serialised. Add return value to allow serialisation.

This fix should be refined, however this bug is currently blocking the
gate so fix it now and improve later.

Change-Id: I583e03b60f7c4be0fa3d65f9dd4ee4e643e4fbab
Closes-bug: #1475716