Remove duplicate app loading
Testing app has be and should be loaded in the FunctionalTest class.
So we should remove the duplicate app loading in the TestCase class.

The config.py file is not used if we remove the duplicate app
loading.So let's remove it.

If we remove the duplicate app loading, TestContainerController
should be modified. The base class of it should be FunctionalTest.

By the way, rename path_get to get_path.

Change-Id: I93d69d9392ecb8ae0303efe14574a6ebe9a82b3c
Closes-Bug: #1478194