Fix setup of tests to remove intermittent failure
The tests in test_periodic can fail if the tests are run in the same
process/thread. The stub bays used in the test were not reset between
tests, leading to intermittent failures in CI. Move the setup of the
stub bays inside the setUp function to assure the stubs are setup
between test functions.

Change-Id: Idb705bc9e5ecb5c798b09367bd546655f874c04e
Closes-Bug: 1476207