Add 'master_addresses' attribute to bay
A bay has an attribute 'node_addresses' that lists
the IP addresses of each minion node. As the support
of HA mode was added, we need a similar attribute,
known as 'master_addresses', that lists the IP address
of each master node. The fix adds 'master_addresses'
attribute to the Bay object and Bay db model. Changes to
add 'master_addresses' column to Bay are also included
in this patch.

Change-Id: I31965a0397bc5c18ea798eb73791d9ef8c209960
Partial-Bug: #1483047