Gate failure due to introduction of new WSME 0.8.0
Many of the changes on gate are failing because of
Bad Request error. The reason for this is the request
accepts some fields whereas the entire dictionary is
getting passed without applying any filter. This
patch solves the problem. A new WSME version 0.8.0 was
introduced today, which seems to be the cause of this
problem.

http://osdir.com/ml/openstack-dev/2015-08/msg01687.html

Change-Id: I6d40889683449995df7bd7ebcdf755991cef43ef
Closes-bug: #1488687