Add version hashes to enforce version updates
Nova uses a map of hashes for important parts of versioned objects. It
keeps a static mapping of previous objects, and generates a map of the
current objects. If these differ, a test fails, enforcing that the hash
of the object needs to be updated. Along with this, it usually means
that the version of the object itself needs to be updated because the
RPC contract of that object is now broken.

oslo_versionedobjects contains a fixture for helping generate and test
these hashes, so this fixture is used to test Magnum's object hashes.

Change-Id: I586f73570dc7b65c83b216d09d11e8dba3c63b2c
Closes-Bug: #1491855