Add UNKNOWN constant to container statuses
The container API would set the status of a container to 'Unknown' if
there was an exception when retrieving the container. This status should
be contained as a constant (as with all other container constants). This
change adds an UNKNOWN constant to the container object module, and
changes the API to use that constant instead of the hard-coded string.

Change-Id: I7cec02e2e8b8619bd2fcba4ba867ec8b21ad5afe
Closes-Bug: #1489996