Minor documentation correction
The glance image list output shown int he docuemntation does not
match the command, which is missing the -v flag. This patch adds
the three additional characters to the doc to make the command
consistent withthe output.

This is a trivial fix that does not require a bug to be tracked.

Change-Id: I4f29dd5d0ac839ed8280695919106a2a87669a68