Fix a 409 failure on bay-update
The error is caused by change of user_token on bay-update. In
particular, when a bay is created, magnum pass user token to heat
as a parameter. Heat in term pass the token to each bay node via
user_data. When a bay is scaled up/down, another token is generated
and passed to heat, which changes the user_data of each bay nodes.
As a result, heat have to replace all the bay nodes to update the
user_data, which cause the error.

Change-Id: If3c3a0d435893d4fa4ed96f7ffd3c414200770f8
Closes-Bug: #1504689