Fix command line in document example
Example in document for Kubernetes Load Balancer.

In the example, the commands to manually associate a floating IP
to the LB VIP uses grep on "TCP" to find the VIP from the neutron
list.  Recently the etcd pool protocol changes to TCP, so the
grep now returns 2 entries instead of just the entry for the LB.
This in turn gives an error in the next command.

Add another grep to filter out the etcd VIP.

Change-Id: I619fa09fb4a5db7bacbc5d73e38016b8cd6404c6
Closes-Bug:  1505780