Create BayModel with provided(VM/BM) server type
In its current state of art, the Baymodel is created
with only "VM" server type. This value is hardcoded in the
code and a tech debt bug was filed to clean it off. Also
this feature can be useful to allow bay model creation with
various server type like virtual machines, baremetal etc.

Client side changes will be provided as part of seperate
patch.

Partial-Bug: #1491003

Change-Id: I8981031d097fc41c94ad509c7b090f0a57a90525