Fix the neutron-lbaas agent config
This patch should fix the functional gate. The issue is that devstack
switch to stevedore aliases for interface_driver configuration [1],
but lbaas doesn't use aliases. This patch fixed the broken config.
This patch should be considered as a temporary workaround. We need
to investigate why devstack doesn't populate the correct config at
the first place.

[1] https://review.openstack.org/#/c/237519/

Closes-Bug: #1512933
Change-Id: I73b9e146ad4cc253e6a04bcadc07f124a7c96df5