Swarm: Add TimeoutStartSec=300 to docker service
The default value of TimeoutStartSec=90, but it will fail on gate.
set TimeoutStartSec=300 to docker service.
It is hard to decide which one is the best choice, I tested on gate, 180 is
fine, some other guys tested on their local setup, said 4 mins, let's
hard code it as 300s first, if docker failed to start at 5 mins, it would be
failed.

see link [1] for reference.
[1] http://man7.org/linux/man-pages/man5/systemd.service.5.html

Co-Authored-By: Kai Qiang Wu <wkqwu@cn.ibm.com>
Partially implements: blueprint swarm-functional-testing
Change-Id: I4fe4f607a2353010092ff565a7189ad3189b0ea9