Add support to set env to a container
Docker run support an option to set environment variables
to a container. container-create in Magnum also needs this.
This will help for the user to create a container with some
environment variables. Also, swarm use some special environments
to scheduler. This patch add support to do this.

Implements: blueprint support-to-set-env-variables

Change-Id: Ifa46eb6d0e528eadca44d265e906c92ee752af43