Fix mesos monitor for handling multiple masters
Magnum is going to enable HA mode for mesos bay, in which there
could be more than one master nodes. One of the master node (leader)
is active and the rest of them is not active. Mesos monitor needs to
identify the leader and pull metrics from the leader only.

Change-Id: Ic7d6cb64964db2758163bd231dec4190302155b2
Partial-Implements: blueprint mesos-multi-master-node