Skeleton for User Guide
Initial empty framework to fill in.

Following up on the discussion at the Tokyo Summit, we are
putting together a User Guide for Magnum.  The initial list of
sections is from the brainstorming session, organized into a
particular sequence that seems reasonable, but this can be
reorganized and extended as the team sees fit.

The intention is to allow contributors to work on all
sections in parallel.

Partially implements: blueprint user-guide
Change-Id: I17ba09f833581bd7fcae9c8a328fc89140b704d2