Skeleton for Troubleshooting Guide
Initial empty framework to fill in

Following up on the discussion at the Tokyo Summit, we are
putting together a Troubleshooting Guide for Magnum.
This is organized as a list of failure symptoms, with pointers
to techniques for troubleshooting.  The initial list of
failure symptoms is from brainstorming at the Summit, but
is expected to grow as more are identified.

This initial framework is not complete by any mean, but is set up
so that multiple contributors can work in parallel.
New details can be added at any time, but current details should
be kept accurate.

Partially implements: blueprint magnum-troubleshooting-guide
Change-Id: I83b9b9ba9c76e608b4adafa525caf92aaaaaf880