Propose Magnum Resource Quota
There are multiple ways to slice an OpenStack cloud. Imposing
quota on these various slices puts a limitation on the amount
of resources that can be consumed by a project which helps to
guarantee "fairness" or fair distribution of resource. If a
particular project needs more resources, the concept of
quota, gives the ability to increase the resource count
on-demand, given that the system constraints are not exceeded.

Change-Id: I6013057d689deb66a371adace96082df747f2b54
Partially-Implements: blueprint resource-quota