Proxy support for Mesos cluster
For OpenStack environment behind a firewall, services such as
google/docker cannot be accessed directly so a proxy server
is needed to provide access.Proxy support is already available
for Kubernetes and Swarm clusters. This series of patches adds
the proxy support for Mesos cluster.

This particular patch covers the Mesos master node.

Implements: blueprint add-proxy-mesos
Change-Id: I1c6e212e1477801d6f0d3c4a48df4f0bba5223d4