Add troubleshooting for network
Cover the scenario when nodes cannot connect to internet

The nodes need to connect to the internet for discovery and
image download.  No network connection is a basic failure
and can happen at any time.  This patch describes some
techniques for debugging.

If you have encountered a scenario not described here, please
mention in the review.

Partially implements: blueprint magnum-troubleshooting-guide
Change-Id: I69ca9d624b960329109fbbed8634646e63aa5118