Troubleshooting Kubernetes networking
Add guide for debugging inter-pods communication

The networking between pods is different and separate from the
neutron network set up for the cluster.  Flannel is the default
network for Kubernetes cluster and Docker is configured to use
Flannel subnet.  If there is a problem, the pods will not be
able to talk to each other.

This section describes techniques for debugging Kubernetes
networking based on Flannel.

Partially implements: blueprint magnum-troubleshooting-guide
Change-Id: Ifaa71b6c8a236041b033da8ea0a8e403ff884594