Adds standardised error messages
Implements the recommendation set out by the API working group for
standardised error messages, available here:
http://specs.openstack.org/openstack/api-wg/guidelines/errors.html

Magnum currently has most of its errors structured by WSME exception
classes, which ideally should be changed to follow API-WG recommendation
too. However, this is outside scope. For now, a workaround is
implemented so that Magnum's API can be locked for v1. When the common
solution to WSME is implemented, most of this patch can be refactored.

Change-Id: I05d2bb60b211d1b18611070217b01e3e418eb79a
Implements: blueprint standardised-error-messages
Signed-off-by: Jamie Hannaford <jamie.hannaford@rackspace.com>