Replace hardcoded eth0 interface in scripts
Make scripts adaptable to different network interface.

Some scripts currently query eth0 directly to get the IP of
the node.  This causes the script to fail if the node uses
a different network interface.  The change passes in the
IP from Heat so that it is not susceptible to the particular
network interface being used.

This change is necessary to use the stock image from Atomic,
since eth0 is not used in this image.

This patch is broken out from the patch:
https://review.openstack.org/#/c/276232/
so that it can proceed independently.

Co-Authored-By: Corey O'Brien <coreypobrien@gmail.com>
Partially-Implements: blueprint atomic23
Change-Id: If8f972d8dabc8304484dfaff8d4e7f1f8755507b