Allow update baymodel's public field even if referenced
Allow update baymodel's public field to be True even if referenced by bays,
User case is one operator wants to share this baymodel to others but doesn't
want to delete the referenced bay.

Please note we only allow to update public to be True case. In case of
True->False, this may lead bays lost baymodel, so don't do this.

Closes-Bug: #1557943
Change-Id: Ia08a2e9611de3559c5cec4eee832bade1f9af09e