Fix bashisms in enable-kube scripts
A few of the scripts in magnum/templates/kubernetes/fragments/ pretend
to be using ISO shell (according to the shebang), but they really are
bash scripts because of 2 small bashisms. This makes lintian (the
Debian package linter) yell. Plus this really would fail with dash
as standard shell implementation for /bin/sh.

This patch fixes the bashisms, rather than the shebang, which is the
best thing to do (so one can choose the shell implementation, and
we're not bound to using bash).

Change-Id: I825a176d34ba985dd85a8bf24d5bcaf1771f6c1a