Add script to install image build dependencies
In order for automated image builds to run, create
this script that can be called from magnum jobs, to
install needed dependencies for diskimage-builder
and fedora-atomic builds.
That script will be run on the VM that will be
spinned to run the image builds, targetted for ubuntu-trusty.

Change-Id: Icf5a2c79e47e632be2ac49bc9ccc528b4f604eb2
Implements blueprint fedora-atomic-image-build